fix(api): tutor name for missions
