+++ api/src/utils/index.js
@@ -29,7 +29,7 @@ const {
   CELLAR_KEYSECRET_SUPPORT,
   PUBLIC_BUCKET_NAME_SUPPORT,
 } = require("../config");
-const { YOUNG_STATUS_PHASE2, SENDINBLUE_TEMPLATES, YOUNG_STATUS, MISSION_STATUS, APPLICATION_STATUS, FILE_STATUS_PHASE1, ROLES } = require("snu-lib");
+const { YOUNG_STATUS_PHASE2, SENDINBLUE_TEMPLATES, YOUNG_STATUS, APPLICATION_STATUS, FILE_STATUS_PHASE1, ROLES } = require("snu-lib");
 
 const { translateFileStatusPhase1 } = require("snu-lib/translation");
 const { SUB_ROLES } = require("snu-lib/roles");
@@ -567,75 +567,6 @@ const getBaseUrl = () => {
   return "http://localhost:8080";
 };
 
-const updateApplicationTutor = async (mission, fromUser = null) => {
-  const applications = await ApplicationModel.find({
-    missionId: mission._id,
-  });
-
-  for (let application of applications) {
-    if (application.tutorId !== mission.tutorId) {
-      const tutor = await ReferentModel.findById(mission.tutorId);
-      application.set({ tutorId: mission.tutorId, tutorName: `${tutor?.firstName} ${tutor?.lastName}` });
-      await application.save({ fromUser });
-    }
-  }
-};
-
-const updateApplicationStatus = async (mission, fromUser = null) => {
-  if (![MISSION_STATUS.CANCEL, MISSION_STATUS.ARCHIVED, MISSION_STATUS.REFUSED].includes(mission.status))
-    return console.log(`no need to update applications, new status for mission ${mission._id} is ${mission.status}`);
-  const applications = await ApplicationModel.find({
-    missionId: mission._id,
-    status: {
-      $in: [
-        APPLICATION_STATUS.WAITING_VALIDATION,
-        APPLICATION_STATUS.WAITING_ACCEPTATION,
-        APPLICATION_STATUS.WAITING_VERIFICATION,
-        // APPLICATION_STATUS.VALIDATED,
-        // APPLICATION_STATUS.IN_PROGRESS,
-      ],
-    },
-  });
-  for (let application of applications) {
-    let cta = `${APP_URL}/phase2`;
-    let statusComment = "";
-    let sendinblueTemplate = "";
-    switch (mission.status) {
-      case MISSION_STATUS.REFUSED:
-        statusComment = "La mission n'est plus disponible.";
-        break;
-      case MISSION_STATUS.CANCEL:
-        statusComment = "La mission a été annulée.";
-        sendinblueTemplate = SENDINBLUE_TEMPLATES.young.MISSION_CANCEL;
-        cta = `${APP_URL}/phase2?utm_campaign=transactionnel+mig+annulee&utm_source=notifauto&utm_medium=mail+261+acceder`;
-        break;
-      case MISSION_STATUS.ARCHIVED:
-        statusComment = "La mission a été archivée.";
-        sendinblueTemplate = SENDINBLUE_TEMPLATES.young.MISSION_ARCHIVED;
-        break;
-    }
-    application.set({ status: APPLICATION_STATUS.CANCEL, statusComment });
-    await application.save({ fromUser });
-
-    // ! Should update contract too if it exists
-
-    if (sendinblueTemplate) {
-      const young = await YoungModel.findById(application.youngId);
-      let cc = getCcOfYoung({ template: sendinblueTemplate, young });
-
-      await sendTemplate(sendinblueTemplate, {
-        emailTo: [{ name: `${application.youngFirstName} ${application.youngLastName}`, email: application.youngEmail }],
-        params: {
-          cta,
-          missionName: mission.name,
-          message: mission.statusComment,
-        },
-        cc,
-      });
-    }
-  }
-};
-
 const getCcOfYoung = ({ template, young }) => {
   if (!young || !template) return [];
   let cc = [];
@@ -905,8 +836,6 @@ module.exports = {
   STEPS,
   STEPS2023,
   STEPS2023REINSCRIPTION,
-  updateApplicationStatus,
-  updateApplicationTutor,
   FILE_STATUS_PHASE1,
   translateFileStatusPhase1,
   getCcOfYoung,
